refactor: 원하는 날짜 형식으로 변환하는 메소드를 Extension으로 수정
